% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_add_input}
\alias{es_add_input}
\title{Adds a shiny input object to the UI.}
\usage{
es_add_input(shinyfunction, tab = "Output", box = "Result", ...)
}
\arguments{
\item{shinyfunction}{a shiny function to call to create the object}

\item{tab}{tab to show the object in (new name creates tab)}

\item{box}{box in the view area to show object in (new name creates box)}

\item{...}{the arguments to deliver to the shiny function (\code{inputId} and \code{value} create
an entry in the \code{input$inputId} list useful in console mode)}
}
\description{
Useful for inputs (e.g. \code{\link{textInput}}).
The input variables can be used in shiny style (\code{input$variable}) in the output functions.
Do not use this function to create output objects, as they are registered in a different way.
The specialized functions (e.g. \code{\link{es_renderPlot}}) can be used for that.
}
