% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_add_object}
\alias{es_add_object}
\title{Adds a shiny object to the UI. Useful for inputs (e.g. \link{\code{shiny::textInput}}).
The input variables can be used in shiny style (\code{input$variable}) in the output functions.
Do not use it for output objects, as they are registered in a different way.
The specialized functions (e.g. \link{\code{easyshiny::add_plot}}) can be used for that.}
\usage{
es_add_object(object, tab = "output", box = "objects")
}
\arguments{
\item{object}{a shiny object to insert}

\item{tab}{tab to show the object in (new name creates tab)}

\item{box}{box in the view area to show object in (new name creates box)}
}
\description{
Adds a shiny object to the UI. Useful for inputs (e.g. \link{\code{shiny::textInput}}).
The input variables can be used in shiny style (\code{input$variable}) in the output functions.
Do not use it for output objects, as they are registered in a different way.
The specialized functions (e.g. \link{\code{easyshiny::add_plot}}) can be used for that.
}
