% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_output.R
\name{es_add_output}
\alias{es_add_output}
\title{Add Output Object}
\usage{
es_add_output(render_call, ui_function, tab = NULL, box = NULL, ...)
}
\arguments{
\item{render_call}{a call to a shiny render function, that generates an output (like \code{\link{renderPlot}}),
the corresponding output function will be automatically chosen}

\item{ui_function}{a shiny UI function, that has to be corresponding to the function used in \code{render_call}}

\item{tab}{tab to show the plot in (default: 'Output', new name creates tab)}

\item{box}{box in the view area to show plot in (default: 'Result', new name creates box)}

\item{...}{the parameters, that should be given to the output function}
}
\description{
Adds an output object from shiny to the Easy Shiny app and places it into a specific tab and box
  after the previously inserted elements. All the wrapper functions pass all their arguments to
  \code{es_add_output} (useful for specifying tabs and boxes).
}
