% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_add_setname}
\alias{es_add_setname}
\title{Set the setnames}
\usage{
es_add_setname(data, setnames)
}
\arguments{
\item{data}{a \code{\link[tibble]{tibble}} without setnames, but containing a column named 'n' for grouping}

\item{setnames}{a vector of setnames at least as long as n in the \code{\link[tibble]{tibble}}}
}
\value{
the altered \code{\link[tibble]{tibble}}
}
\description{
Add setnames to a \code{\link[tibble]{tibble}} using the group identifier n.
}
