% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_add_static}
\alias{es_add_static}
\title{Add Static Component}
\usage{
es_add_static(static_call, tab = "Output", box = "Result")
}
\arguments{
\item{static_call}{a call to a \code{\link[shiny]{tags}} element or a corresponding wrapper}

\item{tab}{the tab to place the static element in}

\item{box}{the box to place the static element in}
}
\description{
Adds a static shiny component to the easyshiny app.
  The object can be composed of different nested shiny calls and will be placed in the given tab and box.
}
