% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_build_ui.R
\name{es_build_objects}
\alias{es_build_objects}
\title{Build Objects}
\usage{
es_build_objects(objects)
}
\arguments{
\item{objects}{a matrix of object descriptions (from \code{globalenv$visuals})}
}
\value{
a list of shiny \code{\link[shinydashboard]{box}}es
}
\description{
Builds the objects for the UI. Shall be used inside a \code{\link[shinydashboard]{box}}.
}
