% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_start.R
\name{es_build_ui}
\alias{es_build_ui}
\title{Builds the UI skeleton for the app.}
\usage{
es_build_ui(title)
}
\value{
html + js code with the needed shiny directives
}
\description{
Builds the UI skeleton for the app.
}
