% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_filechecks.R
\name{es_file_present}
\alias{es_file_present}
\title{Check for File}
\usage{
es_file_present(filesets, setnumber, filename)
}
\arguments{
\item{filesets}{the fileset to search in}

\item{setnumber}{the number of the set to search}

\item{filename}{the name of the desired file}
}
\value{
a boolean value (\code{TRUE} if the file exists, else \code{FALSE})
}
\description{
Checks for the existence of a file in a fileset.
}
