% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_filechecks.R
\name{es_fileset_present}
\alias{es_fileset_present}
\title{Check for Fileset}
\usage{
es_fileset_present(filesets, setnumber)
}
\arguments{
\item{filesets}{the fileset list}

\item{setnumber}{the number of the set to check}
}
\value{
a boolean value (TRUE if the fileset exists, else FALSE)
}
\description{
Checks if a fileset has been uploaded. Used when dynamically loading the upload fields.
}
