% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_init.R
\name{es_init}
\alias{es_init}
\title{Initializes and resets all datastructures used by easyshiny.}
\usage{
es_init(local_folder = NULL)
}
\arguments{
\item{local_folder}{a folder on the computer running the server, containing
input files to test the easyshiny program locally (console mode)}
}
\description{
Initializes and resets all datastructures used by easyshiny.
}
