% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_reactive.R
\name{es_reactive}
\alias{es_reactive}
\title{Create Reactive Value}
\usage{
es_reactive(id, expression)
}
\arguments{
\item{id}{the name of the reactive value}

\item{expression}{the expression to evaluate for the value}
}
\description{
Makes a calue usable for calculations.
}
