% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_read}
\alias{es_read}
\title{Create Filereader}
\usage{
es_read(filename, readerId = NULL)
}
\arguments{
\item{filename}{the name of the file to read}

\item{readerId}{a name for the reader function, that can be used in the plots and console mode (default: derives the name from the filename, just cuts the ending)}
}
\description{
Makes a file usable for user input. The file can then be supplied by the user at runtime and
is provided as function under the given name for visuals creation. The function is also created
in the global environment for usage in console mode.

Understands:

* .csv

* .mds
}
