% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_read_filesets}
\alias{es_read_filesets}
\title{Read fileset}
\usage{
es_read_filesets(folders)
}
\arguments{
\item{folders}{an array of paths to folders containing the data to use}
}
\value{
a dataframe in shiny \code{\link{fileInput}} format or \code{NULL} if the folder can not be found
}
\description{
Reads a fileset from folder (shiny \code{\link{fileInput}} format)
}
