% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add_output.R
\name{es_renderPlot}
\alias{es_renderPlot}
\alias{es_renderDataTable}
\alias{es_renderImage}
\alias{es_renderPrint}
\alias{es_renderTable}
\alias{es_renderText}
\alias{es_renderUI}
\title{Add Output Object}
\usage{
es_renderPlot(expr, tab = NULL, box = NULL, ...)

es_renderDataTable(expr, tab = NULL, box = NULL, ...)

es_renderImage(expr, tab = NULL, box = NULL, ...)

es_renderPrint(expr, tab = NULL, box = NULL, ...)

es_renderTable(expr, tab = NULL, box = NULL, ...)

es_renderText(expr, tab = NULL, box = NULL, ...)

es_renderUI(expr, tab = NULL, box = NULL, ...)
}
\arguments{
\item{expr}{the expression to pass to the outputFunction (should be generating the correct object type)}

\item{tab}{tab to show the plot in (default: 'Output', new name creates tab)}

\item{box}{box in the view area to show plot in (default: 'Result', new name creates box)}

\item{...}{the parameters, that should be given to the output function}
}
\description{
Adds an output object from shiny to the Easy Shiny app and places it into a specific tab and box. All the
  wrapper functions pass all their arguments to \code{es_add_output} (useful for specifying tabs and boxes).
}
\section{Functions}{
\itemize{
\item \code{es_renderDataTable}: adds a shiny datatable to the app

\item \code{es_renderImage}: adds a shiny image to the app

\item \code{es_renderPrint}: prints the results in the app

\item \code{es_renderTable}: adds a shiny table to the app

\item \code{es_renderText}: adds a shiny text to the app

\item \code{es_renderUI}: adds shiny UI elements to the app
}}

