% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_start.R
\name{es_start}
\alias{es_start}
\title{Start Easyshiny App.}
\usage{
es_start(title = "Easy Shiny Project")
}
\arguments{
\item{title}{The title to insert in the main menu.}
}
\value{
The output from \code{\link{shinyApp}}.
}
\description{
Create the Shiny server and UI code and start the server.
The Shiny project will contain all previously added data and visualization.
}
