% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_summaries_to_html}
\alias{es_summaries_to_html}
\title{Convert summaries}
\usage{
es_summaries_to_html(summaries)
}
\arguments{
\item{summaries}{a list of summary objects}
}
\value{
html: a paragraph for every summary containing the computed values
}
\description{
Converts a list containing output from \code{\link{summary}} into a html format usable by shiny.
}
