% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_code.R
\name{make_code}
\alias{make_code}
\title{Generate code files required for shiny app (one dataset)}
\usage{
make_code(
  shiny.title = "App",
  shiny.prefix = "sc1",
  shiny.dir = "app",
  enableSubset = TRUE,
  defPtSiz = 1.25,
  theme = "flatly",
  tabs = c("civge", "civci", "gevge", "gem", "gec", "vio", "pro", "hea"),
  about = TRUE,
  font = "Lato",
  ganalytics = NA
)
}
\arguments{
\item{shiny.title}{title for shiny app}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{enableSubset}{specify whether to enable "Toggle to subset cells"
functionality in the shiny app. Default is to enable this functionality}

\item{defPtSiz}{specify default point size for single cells. For example, a
smaller size can be used if you have many cells in your dataset}

\item{theme}{Bootsrap theme}

\item{tabs}{Vector of tab names to include}

\item{about}{Should about page be added as a tab?}

\item{font}{Google font for plots. Defaults to "Lato".}

\item{ganalytics}{Google analytics tracking ID (e.g. "UA-123456789-0")}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing only one dataset. In
particular, two R scripts will be generated, namely \code{server.R} and
\code{ui.R}. If users want to include multiple dataset in one shiny app,
please use \code{make_code_multi()} instead. Note that both
\code{make_file} and \code{make_code} functions are ran when
running the wrapper function \code{make_app}.
}
\author{
John F. Ouyang
}
