% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyCodes.R
\name{makeShinyCodes}
\alias{makeShinyCodes}
\title{Generate code files required for shiny app (one dataset)}
\usage{
makeShinyCodes(
  shiny.title,
  shiny.prefix,
  shiny.dir,
  enableSubset = TRUE,
  defPtSiz = 1.25,
  theme = "flatly",
  tabs = c(1, 2, 3, 4, 5, 6, 7),
  about = TRUE,
  ganalytics = NA,
  extra_css = FALSE
)
}
\arguments{
\item{shiny.title}{title for shiny app}

\item{shiny.prefix}{specify file prefix}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{enableSubset}{specify whether to enable "Toggle to subset cells" 
functionality in the shiny app. Default is to enable this functionality}

\item{defPtSiz}{specify default point size for single cells. For example, a 
smaller size can be used if you have many cells in your dataset}

\item{theme}{Bootsrap theme}

\item{tabs}{Vector of tab numbers to include}

\item{about}{Should about page be added as a tab?}

\item{ganalytics}{Google analytics tracking ID (e.g. "UA-123456789-0")}

\item{extra_css}{Logical indication if additional CSS is to be created.}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing only one dataset. In 
particular, two R scripts will be generated, namely \code{server.R} and 
\code{ui.R}. If users want to include multiple dataset in one shiny app, 
please use \code{makeShinyCodesMulti()} instead. Note that both 
\code{makeShinyFiles} and \code{makeShinyCodes} functions are ran when 
running the wrapper function \code{makeShinyApp}.
}
\author{
John F. Ouyang
}
