% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShinyCodesMulti.R
\name{makeShinyCodesMulti}
\alias{makeShinyCodesMulti}
\title{Generate code files required for shiny app (multi datasets)}
\usage{
makeShinyCodesMulti(
  shiny.title,
  shiny.prefix,
  shiny.headers,
  shiny.dir,
  enableSubset = TRUE,
  defPtSiz = 1.25,
  theme = "flatly",
  tabs = c(1, 2, 3, 4, 5, 6, 7),
  about = TRUE,
  ganalytics = NA
)
}
\arguments{
\item{shiny.title}{specify the overall title for shiny app}

\item{shiny.prefix}{specify file prefix for each dataset. Must match the
prefix used in \code{makeShinyFiles}}

\item{shiny.headers}{specify the tab header names for each dataset. Length
must match that of \code{shiny.prefix}}

\item{shiny.dir}{specify directory to create the shiny app in}

\item{enableSubset}{specify whether to enable "Toggle to subset cells"
functionality in the shiny app. Default is to enable this functionality}

\item{defPtSiz}{specify default point size for single cells. For example, a
smaller size can be used if you have many cells in your dataset. A single
value can be specified to set the point size for all datasets. Otherwise,
users have to specify one value for each dataset}

\item{theme}{Bootsrap theme}

\item{tabs}{Vector of tab numbers to include}

\item{about}{Should about page be added as a tab?}

\item{ganalytics}{Google analytics tracking ID (e.g. "UA-123456789-0")}
}
\value{
server.R and ui.R required for shiny app
}
\description{
Generate code files required for shiny app containing multiple datasets. In
particular, two R scripts will be generated, namely \code{server.R} and
\code{ui.R}. Note that \code{makeShinyFiles} has to be ran prior to
generate the necessary data files for each of the dataset to be included.
The prefix used in \code{makeShinyFiles} have to be then supplied in this
function.
}
\author{
John F. Ouyang
}
