% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_build_ui.R
\name{mat_col}
\alias{mat_col}
\title{helper to extract a column from a matrix}
\usage{
mat_col(mat, colname)
}
\arguments{
\item{mat}{a matrix}

\item{colname}{the name of the column to extract}
}
\value{
a vector with the data from the column
}
\description{
helper to extract a column from a matrix
}
