% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_colour.R
\name{mod_colour}
\alias{mod_colour}
\title{Modify the colour palette for categorical metadata}
\usage{
mod_colour(scConf, m, nc)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{m}{metadata for which to modify the colour palette. Users 
can either use the actual metadata column names or display names. Please 
specify only one metadata}

\item{nc}{character vector of new colour palette}
}
\value{
updated shinycell config data.table
}
\description{
Modify the colour palette for categorical metadata.
}
\examples{
\dontrun{
scConf = mod_colour(scConf, m = "library", nc = c("black", "darkorange", "blue", "red"))
}
}
\author{
John F. Ouyang

Roy Francis
}
