% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_label.R
\name{mod_label}
\alias{mod_label}
\title{Modify the legend labels for categorical metadata}
\usage{
mod_label(scConf, m, new.labels)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{m}{metadata for which to modify the legend labels. Users 
can either use the actual metadata column names or display names. Please 
specify only one metadata}

\item{new.labels}{character vector of new legend labels}
}
\value{
updated shinycell config data.table
}
\description{
Modify the legend labels for categorical metadata.
}
\examples{
\dontrun{
scConf = mod_label(scConf, m = "library", new.labels = c("Fib", "Primed", "Naive", "RSeT"))
}

}
\author{
John F. Ouyang
}
