% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_meta.R
\name{reorder_meta}
\alias{reorder_meta}
\title{Reorder the order in which metadata appear in the shiny app}
\usage{
reorder_meta(scConf, nmo)
}
\arguments{
\item{scConf}{shinycell config data.table}

\item{nmo}{character vector containing new order. All metadata 
names must be included, which can be found at \code{scConf$ID}}
}
\value{
updated shinycell config data.table
}
\description{
Reorder the order in which metadata appear in the dropdown menu in the 
shiny app.
}
\examples{
\dontrun{
scConf = reorder_meta(scConf, scConf$ID[c(1,3,2,4:length(scConf$ID))])
}
}
\author{
John F. Ouyang

Roy Francis
}
