% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_order.R
\name{show_order}
\alias{show_order}
\title{Shows the order in which metadata will be displayed}
\usage{
show_order(scConf)
}
\arguments{
\item{scConf}{shinycell config data.table}
}
\value{
table showing the order in which metadata will be displayed
}
\description{
Shows the order in which metadata will be displayed in the shiny app. This 
helps users to decide if the display order is ok. If not, users can use 
\code{reorder_meta} to change the order in which metadata will be displayed.
}
\examples{
\dontrun{
show_order(scConf)
}

}
\author{
John F. Ouyang
}
