% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wr_load}
\alias{wr_load}
\title{Write code for loading objects for ui.R and server.R}
\usage{
wr_load(prefix, tabs)
}
\arguments{
\item{prefix}{file prefix}

\item{tabs}{Vector of tab names to include}
}
\description{
Write code for loading objects for ui.R and server.R
}
