% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wr_sv_load}
\alias{wr_sv_load}
\title{Write code for loading objects for server.R}
\usage{
wr_sv_load(prefix)
}
\arguments{
\item{prefix}{file prefix}
}
\description{
Write code for loading objects for server.R
}
