% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wr_sv_main}
\alias{wr_sv_main}
\title{Write code for main block of server.R}
\usage{
wr_sv_main(
  prefix,
  subst = "",
  tabs = c("civge", "civci", "gevge", "gem", "gec", "vio", "pro", "hea")
)
}
\arguments{
\item{prefix}{file prefix}

\item{subst}{Conditional}

\item{tabs}{Vector of tab names to include}
}
\description{
Write code for main block of server.R
}
