% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUImain}
\alias{wrUImain}
\title{Write code for main block of ui.R}
\usage{
wrUImain(
  prefix,
  subst = "",
  ptsiz = "1.25",
  tabs = c(1, 2, 3, 4, 5, 6, 7),
  about = TRUE
)
}
\arguments{
\item{prefix}{file prefix}

\item{subst}{Conditional}

\item{ptsiz}{Point size}

\item{tabs}{Vector of tab numbers to include}

\item{about}{Should about page be added as a tab?}
}
\description{
Write code for main block of ui.R
}
