% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wrUItab2}
\alias{wrUItab2}
\title{Write code for tab2 of ui.R}
\usage{
wrUItab2(prefix, subst = "", ptsiz = "1.25")
}
\arguments{
\item{prefix}{file prefix}

\item{subst}{Conditional}

\item{ptsiz}{Point size}
}
\description{
Write code for tab2 of ui.R
}
