% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_add_object}
\alias{es_add_object}
\title{Adds a shiny object to the UI.}
\usage{
es_add_object(shinyfunction, tab = "output", box = "objects", ...)
}
\arguments{
\item{shinyfunction}{a shiny function to call to create the object}

\item{tab}{tab to show the object in (new name creates tab)}

\item{box}{box in the view area to show object in (new name creates box)}

\item{...}{the arguments to deliver to the shiny function (\code{inputId} and \code{value} create
an entry in the \code{input$list} useful for console mode)}
}
\description{
Useful for inputs (e.g. \code{\link{textInput}}).
The input variables can be used in shiny style (\code{input$variable}) in the output functions.
Do not use it for output objects, as they are registered in a different way.
The specialized functions (e.g. \code{\link{es_add_plot}}) can be used for that.
}
