% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_add.R
\name{es_add_output}
\alias{es_add_output}
\title{Add Plot}
\usage{
es_add_output(renderFunction, expr, tab = "Output", box = "Result",
  outputId = NULL, ...)
}
\arguments{
\item{renderFunction}{a shiny render function, that generates an output (like \code{\link{renderPlot}}),
the corresponding output function will be automatically chosen}

\item{expr}{the expression to pass to the outputFunction (should be generating the correct object type)}

\item{tab}{tab to show the plot in (default: 'Output', new name creates tab)}

\item{box}{box in the view area to show plot in (default: 'Result', new name creates box)}

\item{outputId}{(optional) if given, defines a function in global environment, that generates the plot}

\item{...}{the parameters, that should be given to the output function}
}
\description{
Adds a plot to the Easy Shiny app and places it into a specific tab and box.
}
