% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_add_setname}
\alias{es_add_setname}
\title{Add setnames to a data frame using the group identifier n.}
\usage{
es_add_setname(simdata, setnames)
}
\arguments{
\item{simdata}{a dataframe without setnames}

\item{setnames}{a vector of setnames at least as long as n in the data frame}
}
\value{
the altered data frame
}
\description{
Add setnames to a data frame using the group identifier n.
}
