% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_build_ui.R
\name{es_build_objects}
\alias{es_build_objects}
\title{Builds the objects for the UI. To use in a box.}
\usage{
es_build_objects(objects)
}
\arguments{
\item{objects}{a matrix of object descriptions (from the visuals)}
}
\value{
a list of shiny boxes
}
\description{
Builds the objects for the UI. To use in a box.
}
