% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_build_ui.R
\name{es_build_ui}
\alias{es_build_ui}
\title{Build UI}
\usage{
es_build_ui(title, visuals)
}
\arguments{
\item{title}{The title to insert in the main menu.}

\item{visuals}{The visuals to use for the UI (as created in \code{globalenv$visuals})}
}
\value{
html + js code with the needed shiny directives
}
\description{
Builds the UI skeleton for the app.
}
