% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_read}
\alias{es_read}
\title{Makes a file usable for user input. The file can then be supplied by the user at runtime and
is provided under its name for visuals creation. Understands:}
\usage{
es_read(filename)
}
\arguments{
\item{filename}{the name of the file to read}
}
\description{
* .csv
* .mds
}
