% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_read_exfiles}
\alias{es_read_exfiles}
\title{Wrapper for \code{\link{es_read_files}}. To use for reading files without header.}
\usage{
es_read_exfiles(filesets, filename, callback)
}
\arguments{
\item{filesets}{see \code{\link{es_read_files}}}

\item{filename}{see \code{\link{es_read_files}}}

\item{callback}{see \code{\link{es_read_files}}}
}
\value{
see \code{\link{es_read_files}}
}
\description{
Wrapper for \code{\link{es_read_files}}. To use for reading files without header.
}
