% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_read.R
\name{es_read_files}
\alias{es_read_files}
\title{Read file}
\usage{
es_read_files(filesets, filename, prepare = function(data) {     data },
  ...)
}
\arguments{
\item{filesets}{a filesets containing one or more files with the name}

\item{filename}{the name of the file to use}

\item{prepare}{(optional) a function to call on to alter the read \code{\link[tibble]{tibble}}}

\item{...}{arguments to pass on to \code{\link{read.csv}}}
}
\value{
a \code{\link[tibble]{tibble}} containing the read data, \code{NULL} if the file is not found
}
\description{
Read a specific file from multiple sources in the fileset and join to one \code{\link[tibble]{tibble}}.
}
