% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_build_ui.R
\name{mat_col}
\alias{mat_col}
\title{Extract column}
\usage{
mat_col(mat, colname)
}
\arguments{
\item{mat}{a \code{\link{matrix}}}

\item{colname}{the name of the column to extract}
}
\value{
a \code{\link{vector}} with the data from the column
}
\description{
A helper to extract a column from a \code{\link{matrix}}
}
