% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writer.R
\name{wr_ui_main}
\alias{wr_ui_main}
\title{Write code for main block of ui.R}
\usage{
wr_ui_main(
  prefix,
  subst = "",
  ptsiz = "1.25",
  tabs = c("civge", "civci", "gevge", "gem", "gec", "vio", "pro", "hea"),
  about = TRUE
)
}
\arguments{
\item{prefix}{file prefix}

\item{subst}{Conditional}

\item{ptsiz}{Point size}

\item{tabs}{Vector of tab names to include}

\item{about}{Should about page be added as a tab?}
}
\description{
Write code for main block of ui.R
}
